requirements.txt
fastapi
uvicorn
httpx
pydantic
llama-cpp-python